Update SDK and CLI to use new API, and frameworks support

+++ packages/cli/src/commands/dev/watch.ts
@@ -10,13 +10,21 @@ import webpack from 'webpack';
 
 import {createFunctionManifest} from './manifest';
 import {formatWebpackMessages} from '../../shared/format-webpack-messages';
-import {getLocalConfig} from '../../shared/fleet-config';
+import {getFramework, Manifest} from '../../shared/frameworks';
+import {getLocalConfig, FleetConfig} from '../../shared/fleet-config';
 import {reportWebpackWarnings} from '../../shared/webpack-error-utils';
 import report from '../../reporter';
 import store, {Actions} from './store';
 
-function getFleetFunctionConfig(pathDir: string) {
-  const localConfig = getLocalConfig(pathDir);
+function getFleetFunctionConfig(pathDir: string, frameworkManifest?: Manifest) {
+  let localConfig = getLocalConfig(pathDir);
+
+  if (frameworkManifest) {
+    localConfig = {
+      ...(localConfig ?? {}),
+      functions: frameworkManifest.functions,
+    } as FleetConfig;
+  }
 
   if (!localConfig) {
     return report.panic(
@@ -35,11 +43,12 @@ function getFleetFunctionConfig(pathDir: string) {
   return localConfig;
 }
 
-async function createWebpackConfig(
+function getFunctionManifest(
   compiledFunctionsDir: string,
-  pathDir: string
+  pathDir: string,
+  frameworkManifest?: Manifest
 ) {
-  const localConfig = getFleetFunctionConfig(pathDir);
+  const localConfig = getFleetFunctionConfig(pathDir, frameworkManifest);
 
   const manifest = createFunctionManifest(
     pathDir,
@@ -48,8 +57,20 @@ async function createWebpackConfig(
   );
 
   store.dispatch({payload: manifest, type: Actions.ADD_MANIFEST});
+  store.dispatch({payload: localConfig, type: Actions.ADD_LOCAL_CONFIG});
 
-  const env = localConfig.env ?? {};
+  return {
+    manifest,
+    localConfig,
+  };
+}
+
+async function createWebpackConfig(
+  compiledFunctionsDir: string,
+  pathDir: string,
+  entries: Record<string, string>,
+  env: Record<string, string> = {}
+) {
   const nodeEnv = process.env.NODE_ENV || 'development';
 
   const processEnv = Object.keys(env).reduce<Record<string, string>>(
@@ -62,7 +83,7 @@ async function createWebpackConfig(
   );
 
   return {
-    entry: manifest.entries,
+    entry: entries,
     output: {
       path: compiledFunctionsDir,
       filename: '[name].js',
@@ -115,15 +136,19 @@ async function createWebpackConfig(
 }
 
 export async function watch(pathDir: string) {
-  const spinnerStop = report.createSpinner('Compiling Fleet Functions');
+  const frameworkManifest = await getFramework(pathDir);
+
+  const spinnerStop = report.createSpinner(
+    frameworkManifest ? 'Starting the watcher' : 'Compiling Fleet Functions'
+  );
 
   let isFirstBuild = true;
 
+  const fleetFuntionsCacheDir = path.join('.fleet', 'cache', 'functions');
+
   const compiledFunctionsDir = path.join(
     pathDir,
-    '.fleet',
-    'cache',
-    'functions'
+    frameworkManifest?.compiledFunctionsDir ?? fleetFuntionsCacheDir
   );
 
   store.dispatch({
@@ -131,11 +156,23 @@ export async function watch(pathDir: string) {
     type: Actions.ADD_COMPILED_DIR,
   });
 
-  await fs.ensureDir(compiledFunctionsDir);
-  await fs.emptyDir(compiledFunctionsDir);
+  if (!frameworkManifest) {
+    await fs.ensureDir(compiledFunctionsDir);
+    await fs.emptyDir(compiledFunctionsDir);
+  }
 
   try {
-    const config = await createWebpackConfig(compiledFunctionsDir, pathDir);
+    const manifestAndConfig = getFunctionManifest(
+      compiledFunctionsDir,
+      pathDir,
+      frameworkManifest
+    );
+    const config = await createWebpackConfig(
+      compiledFunctionsDir,
+      pathDir,
+      manifestAndConfig.manifest.entries,
+      manifestAndConfig.localConfig.env
+    );
 
     const compilerCallback = (
       err: Error | undefined,
@@ -185,7 +222,11 @@ export async function watch(pathDir: string) {
       }
     };
 
-    let compiler = webpack(config).watch({}, compilerCallback);
+    let compiler: webpack.Watching;
+
+    if (!frameworkManifest) {
+      compiler = webpack(config).watch({}, compilerCallback);
+    }
 
     chokidar
       // Watch the Fleet configuration files to see if a function has been
@@ -200,10 +241,22 @@ export async function watch(pathDir: string) {
           )}`
         );
 
+        const manifestAndConfig = getFunctionManifest(
+          compiledFunctionsDir,
+          pathDir,
+          frameworkManifest
+        );
+
+        if (frameworkManifest) {
+          return;
+        }
+
         compiler.close(async () => {
           const config = await createWebpackConfig(
             compiledFunctionsDir,
-            pathDir
+            pathDir,
+            manifestAndConfig.manifest.entries,
+            manifestAndConfig.localConfig.env
           );
           compiler = webpack(config).watch({}, compilerCallback);
         });
